+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - [#51](https://github.com/CallistoLabsNYC/samsa/issues/51) Add create and delete topic protocol support
+- [#32](https://github.com/CallistoLabsNYC/samsa/issues/32) Manage Redpanda transforms via admin api
 
 ### Fixed
 - [#45](https://github.com/CallistoLabsNYC/samsa/issues/45) Cannot perform ConsumerGroupBuilder::build() via a reference